feat: refactor telemetry
