+++ pkg/constants.go
@@ -1,3 +1,4 @@
 package pkg
 
 const SegmentIOWriteKey = "0gAYkX5RV3vt7s4pqCOOsDb6WHPLT30M"
+const MetricHealth = "kubefirst.gh.mgmt.kubefirst.com.healthy"
